+++ rust/highest_averages/webster.rs
@@ -9,8 +9,8 @@ impl Allocate for WebsterSainteLague {
         total_seats: u32,
         n_parties: usize,
     ) -> AllocationResult {
-        fn quotient(original_votes: u64, n_seats_won: u32) -> u64 {
-            original_votes / (2 * n_seats_won as u64 + 1)
+        fn quotient(original_votes: u32, n_seats_won: u32) -> u32 {
+            original_votes / (2 * n_seats_won + 1)
         }
         allocate_highest_average(quotient, total_seats, &ballots, n_parties)
     }
